Basic code that connects client to server
There are two libraries imported: smtplib and SSL. The mail server used is smtp gmail. Then ehlo method identifies ourselves with the mail server we are using. Then we encrypt it using starttls method. Then ehlo again re-identifies ourselves as a encrypted connection. And when we are connected, we can login to our mail server by passing our username and password to server.login method. Then to check the working of our code we are going to print a statement. If the SMTP client is not able to establish a connection between the client and server then in the except block SMTPException is thrown.