Send_text_message
In this code we are importing: SMTP from smtplib, os module and EmailMessage class. The functions that the OS module provides allows you to interface with the underlying operating system that Python is running on – be that Windows or Mac. The central class in the email package is the EmailMessage class, imported from the email.message module. It is the base class for the email object model. EmailMessage provides the core functionality for setting and querying header fields, for accessing message bodies, and for creating or modifying structured messages.